+++ src/helm/benchmark/static/benchmarking.js
@@ -124,6 +124,44 @@ $(function () {
     return $table;
   }
 
+  function renderPlots() {
+    const container = $('<div>', {class: "container"});
+    const links = $('<div>');
+    container.append(links);
+    const tableLinks = [];
+
+    function renderPlot(name, title) {
+        const plot = $('<div>', {class: "plot"});
+        const caption = $('<div>', {class: "plot-caption"}).append(plotCaptions[name]);
+
+        plot.append($('<h3>').append($('<a>', {id: title}).append(title)));
+        plot.append(caption);
+        plot.append($('<img>', {src: plotUrl(suite, name), class: "img-fluid"}));
+        container.append(plot);
+        tableLinks.push($('<a>', {href: '#' + title}).append(title));
+    }
+
+    renderPlot("generic_summary", "Metric spread for core scenarios");
+    renderPlot("model_ranking_all", "Head-to-head win rate per each model");
+
+    renderPlot("accuracy_v_x", "Accuracy as a function of other metrics");
+    renderPlot("metric_correlation", "Correlation between metrics");
+
+    renderPlot("accuracy_v_access", "Accuracy as a function of model access");
+    renderPlot("accuracy_over_num_parameters", "Accuracy across model sizes");
+    renderPlot("accuracy_over_release_date", "Accuracy over time");
+    renderPlot("accuracy_over_the_pile_perplexity", "Accuracy as a function of The Pile perplexity");
+
+    renderPlot("targeted_evals", "Targeted evaluations");
+
+    renderPlot("in_context_ablations", "Number of in-context examples ablation");
+    renderPlot("mc_ablations", "Multiple-choice adaptation ablation");
+
+    links.append(renderItems(tableLinks));
+
+    return container;
+  }
+
   function renderRunsOverview(runSpecs) {
     let query = '';
     const $search = $('<input>', {type: 'text', size: 40, placeholder: 'Enter regex query (enter to open all)'});
@@ -1172,6 +1210,11 @@ $(function () {
       // Models
       $main.empty()
       $main.append(renderHeader('Scenarios', renderScenarios()));
+      refreshHashLocation();
+    } else if (urlParams.plots) {
+      // Plots
+      $main.empty()
+      $main.append(renderHeader('Plots', renderPlots()));
       refreshHashLocation();
     } else if (urlParams.runSpec || urlParams.runSpecs || urlParams.runSpecRegex) {
       // Predictions for a set of run specs (matching a regular expression)