+++ src/helm/proxy/services/remote_service.py
@@ -7,6 +7,7 @@ from typing import Any, List, Optional
 
 from helm.common.authentication import Authentication
 from helm.common.moderations_api_request import ModerationAPIRequest, ModerationAPIRequestResult
+from helm.common.critique_request import CritiqueRequest, CritiqueRequestResult
 from helm.common.perspective_api_request import PerspectiveAPIRequest, PerspectiveAPIRequestResult
 from helm.common.tokenization_request import (
     TokenizationRequest,
@@ -97,6 +98,9 @@ class RemoteService(Service):
         response = requests.get(f"{self.base_url}/api/moderation?{urllib.parse.urlencode(params)}").json()
         RemoteService._check_response(response, request_json)
         return from_dict(ModerationAPIRequestResult, response)
+
+    def make_critique_request(self, auth: Authentication, request: CritiqueRequest) -> CritiqueRequestResult:
+        raise NotImplementedError("make_critique_request is not supported by RemoteServer")
 
     def create_account(self, auth: Authentication) -> Account:
         data = {"auth": json.dumps(asdict(auth))}